Elevator card can be put in the motherboard.
Players can remove the elevator card too, and this will change the
display message shown in the motherboard's screen.
Added a command to add items to the inventory. Use add_item and pass the
script_name of the item you want to add.
FIX: AudioManager no longer duplicates entries in its cues arrays. Added
a temporary button to the Audio tab to remove duplicate entries.
--- Ready to make the elevator, and its app, to work ---