Updated ElevatorPuzzle to look like and app
Now it shows a map of the building, the links between floors and each
button and a button to close the app (not working yet).
The elevator app CAN'T be opened using its icon. I'll do that tomorrow.
Clickable now has a direct reference to the PopochiuRoom where to which
it belongs. This is set by Popochiu when the Room loads.
Updated the disable state style for the "edit" button in the Main menu.
=== Ready to work on more functionality for the Elevator app and in
making the other two apps to show ===