Created theme for OS.
At the moment is being used only in the Register App, but it will be the
default theme for all the OS interfaces in some PCs.
The created theme defines class overwrites for HScroll, OptionButton and
PopupMenu (Radio buttons included).
Created Resource > Data.gd to store character names and arcanes
temporarily.