Added fortune teller notes and progress in puzzles panel.
Documents now emit a signal when closing them to trigger actions after that.
Documents with a single page doesn't show page numbers anymore.
Interaction with PropSecretCompartiment varies depending on the state of the game.
Fortune teller notes open when interacting with the secret compartiment open. After closing the notes, the Registry USB is added to the inventory.
FIX: Documents were not closable with documents with a single page.