Improved how Props z-index is modified when PC moves through the room.
Now it takes into account the baseline of the Props instead of its order in the tree.
Created an Area2D that prevents closing an Overlay2D when players click the background of the panel.
Updated the spritesheet for the musicbox cabinet in the Policeman room to show both drawers open.
Now the USB is added to the inventory after solving the musicbox puzzle in the Policeman room.
Updated the "wait" cursor.
Updated the type of cursor used in some doors and Hotspots that make players move from one room to another.