Master key description in inventory shows right click description.
PC walks to computer when using the USB and the elevator card.
Players can go to the basement by the elevator will go back to the previous floor because the basement is not ready yet.
Updated the min_size for all inventory items and removed the SIZE_EXPAND flag because that was making the inventory look weird when player has only two items.
Added functions to Popochiu to know if a room exists and to make the transition layer play a transition.