Created ADNAnalyzer popup with behavior.
Updated the textures for the two ADNpicker states.
Updated the texture for the PC in the Doctor's room (a.k.a. ADN Analyzer).
Added a Sprite to the PC in the Doctor's room to show if the picker is or not plugged in the thing with or without an ADN sample.
The ADNAnalyzer opens with no picker, an empty picker or a full picker. For the two former cases the PC says something after closing the popup.