Added sprites to show the ADN analysis results.
Created a shader (thanks @agar3s) to tint the bug parts that will appear as part of the ADN analysis result.
Updated the flow of showing the analysis results in order to show the ADN string once the animations have finished.
The ADNAnalyzer can't be closed until it shows a closing state.