Created Painting popup.
All paintings open the Painting popup, and the hidden layer if the Magic glasses were used to interact with the painting.
The same occurs if the magic glasses are used on the Painting popup.
Created Password script in Painting to encrypt the order in which the pieces should be readed.