Adn analyzer, Paintings and Tarot (#14)
* Created ADNAnalyzer popup with behavior.

Updated the textures for the two ADNpicker states.
Updated the texture for the PC in the Doctor's room (a.k.a. ADN Analyzer).
Added a Sprite to the PC in the Doctor's room to show if the picker is or not plugged in the thing with or without an ADN sample.
The ADNAnalyzer opens with no picker, an empty picker or a full picker. For the two former cases the PC says something after closing the popup.

* Added sprites to show the ADN analysis results.

Created a shader (thanks @agar3s) to tint the bug parts that will appear as part of the ADN analysis result.
Updated the flow of showing the analysis results in order to show the ADN string once the animations have finished.
The ADNAnalyzer can't be closed until it shows a closing state.

* Created Painting popup.

All paintings open the Painting popup, and the hidden layer if the Magic glasses were used to interact with the painting.
The same occurs if the magic glasses are used on the Painting popup.
Created Password script in Painting to encrypt the order in which the pieces should be readed.

* Added code to paint the password in painting.

This depends on a Dictionary that defines in which position each symbol should go.
The Dictionary can be setup in the Painting's Inspector property: code.
Defined a password setup for the Fortune teller painting.

* Created images for green book, blue book and sheets.

This are for the Doctor room.
Updated the node that shows documents' texts to a RichTextLabel.
FIX: The documents background image does not move after the right or exit buttons appear.

* Created Tarot popup