Hidden objects can move following a Path2D.
Added a baseline property to HiddenObject so it makes easier to define the point that Godot must use for sorting the nodes.
Fixed a bug on how objects behind were handled when clicking other objects: now it sets the object behind as the current object only for objects that trigger the 'clicked' signal.