Created Window popup for Writer's window.
Here players can clic on an Area2D to choose a point and after that the jumping meter appears so players can choose the height of the jump.
Created temporary background image for this popup.
No Overlay2DBlocker can trigger a signal when registering a click so other scripts can listen to that and do something.
Hidden minigame can be opened with the Hidden computer in Writer's room.
Hidden objects that move by themselves only start moving when calling their move() public method. This will be useful to put things in action only when a level starts.